+++ packages/core/docs/changelog.md
@@ -1,5 +1,5 @@
 # @twin.org/core - Changelog
 
-## 0.0.1-next.3
+## 0.0.1-next.4
 
 - Initial Release